daemon: swap daemonizing and provider instantiation
With the persistentjabber provider we now have a plugin, that spawns a
thread. As is widely known threading and forking doesn't go well when
mixed. As it happens a provider ctor would create a thread and the
daemonize would kill it. Not what we want. The solution now is first
daemonize, then instantiate.