convert to logging module
This commit should not result in user visible changes. It just changes
the backend for logging and enables adding logging to providers.

Due to a combined bug in the logging module and the pyinotify module we
have a gross hack at the top of pynotifydaemon. CPython trunk has fixed
this.